refactor(productdetail): 상품의 상세 화면에서 다른 화면으로 넘어가는 행동을 수행하는 클래스 ProductDetailNavigator 생성
