+++ app/src/main/java/woowacourse/shopping/productdetail/ProductDetailNavigator.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.productdetail
+
+import woowacourse.shopping.model.ProductUiModel
+
+interface ProductDetailNavigator {
+
+    fun navigateToCartView()
+
+    fun navigateToProductDetailView(product: ProductUiModel)
+}
